backend/requirements.txt
fastapi
uvicorn
gitpython
openai
requests
python-dotenv
python-multipart   
